+++ packages/dart_super/test/src/injection_test.dart
@@ -9,7 +9,6 @@ void main() {
     setUp(() {
       Injection.activate(
         testMode: true,
-        autoDispose: true,
         mocks: [],
         enableLog: true,
       );
@@ -19,22 +18,17 @@ void main() {
 
     test('create() should register a singleton instance', () {
       final instance = MockRx();
-      final lazyInstance = MockController();
 
-      Injection.create<MockRx>(instance);
+      Injection.create<MockRx>(instance, true);
 
       expect(Injection.of<MockRx>(), instance);
-
-      Injection.create<MockController>(lazyInstance, lazy: true);
-
-      expect(Injection.of<MockController>(), lazyInstance);
     });
 
     test('create() should throw an error if SuperApp is not found', () {
       Injection.deactivate();
 
       expect(
-        () => Injection.create<String>('example'),
+        () => Injection.create<String>('example', true),
         throwsA(isA<StateError>()),
       );
     });
@@ -42,7 +36,7 @@ void main() {
     test('of() should retrieve an instance from the manager', () {
       const instance = 'example';
 
-      Injection.create<String>(instance);
+      Injection.create<String>(instance, true);
 
       expect(Injection.of<String>(), instance);
     });
@@ -50,7 +44,7 @@ void main() {
     test('of() should invoke enable() on SuperController instances', () {
       final instance = MockController();
 
-      Injection.create<MockController>(instance);
+      Injection.create<MockController>(instance, true);
 
       expect(instance.enableCalled, false);
 
@@ -59,18 +53,10 @@ void main() {
       expect(instance.enableCalled, true);
     });
 
-    test('of() should register and retrieve lazy instances', () {
-      const lazyInstance = 10;
-
-      Injection.create<int>(lazyInstance, lazy: true);
-
-      expect(Injection.of<int>(), lazyInstance);
-    });
-
     test('of() should recursively register and retrieve instances', () {
       final instance = MockController();
 
-      Injection.create<MockController>(instance);
+      Injection.create<MockController>(instance, true);
 
       final mockInstance = Injection.of<MockController>();
 
@@ -87,23 +73,23 @@ void main() {
     test('init() should retrieve an instance or create a new one', () {
       const instance = 'example';
 
-      Injection.create<String>(instance);
+      Injection.create<String>(instance, true);
       // String key already exists
-      expect(Injection.init<String>('other'), instance);
-      expect(Injection.init<int>(10), 10);
+      expect(Injection.init<String>('other', true), instance);
+      expect(Injection.init<int>(10, true), 10);
     });
 
     test('init() should create a new instance if it does not exist', () {
       const instance = 'example';
 
-      expect(Injection.init<String>(instance), instance);
+      expect(Injection.init<String>(instance, true), instance);
       expect(Injection.of<String>(), instance);
     });
 
     test('delete() should delete an instance from the manager', () {
       const instance = 'example';
 
-      Injection.create<String>(instance);
+      Injection.create<String>(instance, true);
 
       expect(Injection.of<String>(), instance);
 
@@ -115,7 +101,7 @@ void main() {
     test('delete() should disable SuperController instances', () {
       final controller = MockController();
 
-      Injection.create<MockController>(controller);
+      Injection.create<MockController>(controller, true);
 
       expect(controller.disableCalled, false);
 
@@ -127,7 +113,7 @@ void main() {
     test('delete() should dispose Rx instances', () {
       final rxInstance = MockRx();
 
-      Injection.create<Rx<dynamic>>(rxInstance);
+      Injection.create<Rx<dynamic>>(rxInstance, true);
 
       expect(rxInstance.disposeCalled, false);
 
@@ -137,9 +123,9 @@ void main() {
     });
 
     test('deleteAll() should delete all instances from the manager', () {
-      Injection.create<String>('example');
-      Injection.create<int>(42);
-      Injection.create<Rx<dynamic>>(MockRx());
+      Injection.create<String>('example', true);
+      Injection.create<int>(42, true);
+      Injection.create<Rx<dynamic>>(MockRx(), true);
 
       expect(Injection.of<String>(), 'example');
       expect(Injection.of<int>(), 42);
@@ -165,8 +151,8 @@ void main() {
       final controller1 = MockController();
       final controller2 = MockController();
 
-      Injection.create<MockController>(controller1);
-      Injection.create<SuperController>(controller2);
+      Injection.create<MockController>(controller1, true);
+      Injection.create<SuperController>(controller2, true);
 
       expect(controller1.disableCalled, false);
       expect(controller2.disableCalled, false);
@@ -181,8 +167,8 @@ void main() {
       final rxInstance1 = MockRx();
       final rxInstance2 = MockRx();
 
-      Injection.create<MockRx>(rxInstance1);
-      Injection.create<Rx<dynamic>>(rxInstance2);
+      Injection.create<MockRx>(rxInstance1, true);
+      Injection.create<Rx<dynamic>>(rxInstance2, true);
 
       expect(rxInstance1.disposeCalled, false);
       expect(rxInstance2.disposeCalled, false);