+++ packages/flutter_super/lib/src/core/globals.dart
@@ -1,3 +1,0 @@
-/// Map that stores the callbacks registered under an ID
-/// from SuperBuilders.
-final Map<String, List<void Function()>> rebuildMap = {};