+++ packages/flutter_super/lib/src/widgets/super_builder.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_super/flutter_super.dart';
-import 'package:flutter_super/src/core/globals.dart';
 import 'package:flutter_super/src/core/typedefs.dart';
 
 /// A stateful widget that rebuilds its child widget when the [Rx]
@@ -55,12 +54,8 @@ class SuperBuilder extends StatefulWidget {
   const SuperBuilder({
     required this.builder,
     this.buildWhen,
-    this.id,
     super.key,
   });
-
-  /// Marks the widget to be rebuilt using the SuperController rebuild method.
-  final String? id;
 
   /// Called every time the [Rx] object changes state.
   final RxBuilder builder;
@@ -77,13 +72,11 @@ class SuperBuilder extends StatefulWidget {
 
 class _SuperBuilderState extends State<SuperBuilder> {
   RxMerge<dynamic>? _rx;
-  String? id;
 
   @override
   void dispose() {
     // Remove the listener from the rx
     _rx?.removeListener(_handleChange);
-    rebuildMap[id]?.remove(_handleChange);
     super.dispose();
   }
 
@@ -92,28 +85,15 @@ class _SuperBuilderState extends State<SuperBuilder> {
     setState(() {});
   }
 
-  void _init() {
-    id = widget.id;
-    try {
-      _rx = RxListener.listenedRx();
-      _rx?.addListener(_handleChange);
-    } catch (e) {
-      if (id == null) rethrow;
-    }
-    if (id == null) return;
-    if (rebuildMap[id] == null) rebuildMap[id!] = [];
-    rebuildMap[id]?.add(_handleChange);
-  }
-
   @override
   Widget build(BuildContext context) {
     _rx?.removeListener(_handleChange);
-    rebuildMap[id]?.remove(_handleChange);
 
     RxListener.listen();
     final child = widget.builder(context);
 
-    _init();
+    _rx = RxListener.listenedRx();
+    _rx?.addListener(_handleChange);
     return child;
   }
 }